fff-dashboard/airflow/requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.7
sqlalchemy==1.4.53
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
lightgbm==4.1.0
numpy==1.24.3
holidays==0.30
scikit-learn==1.3.0
